chore: validations added

+++ src/functions/deleteTask.ts
@@ -1,18 +1,23 @@
 import { APIGatewayProxyHandler } from 'aws-lambda'
 import AWS from 'aws-sdk'
 
-export const handler: APIGatewayProxyHandler = async (event) => {
-  const { id } = event.pathParameters as { id: string }
+import { getRespond } from '../utils/api';
 
-  const dynamodb = new AWS.DynamoDB.DocumentClient();
+export const handler: APIGatewayProxyHandler = async ({ pathParameters }) => {
+  try {
+    const { id } = pathParameters as { id: string }
 
-  await dynamodb.delete({
-    TableName: 'TaskTable',
-    Key: { id },
-  }).promise()
+    if (!id) return getRespond({ statusCode: 400, message: 'id is required' })
 
-  return {
-    statusCode: 200,
-    body: JSON.stringify({ message: 'Task deleted successfully' }),
-  };
+    const dynamodb = new AWS.DynamoDB.DocumentClient();
+
+    await dynamodb.delete({
+      TableName: 'TaskTable',
+      Key: { id },
+    }).promise()
+
+    return getRespond({ statusCode: 200, message: 'Task deleted successfully' })
+  } catch (error) {
+    return getRespond({ statusCode: 500, message: error.message })
+  }
 };
